+++ src/resources/views-elfinder/filepicker.php
@@ -35,6 +35,7 @@
         $().ready(function () {
             var elf = $('#elfinder').elfinder({
                 // set your elFinder options here
+                resizable: false,
                 <?php if ($locale) {
         ?>
                     lang: '<?= $locale ?>', // locale
@@ -85,11 +86,17 @@
                 }
             }).elfinder('instance');
         });
+        $(window).resize(function(){
+            var h = ($(window).height());
+            if($('#elfinder').height() != h){
+                $('#elfinder').height(h).resize();
+            }
+        });
     </script>
 
 
 </head>
-<body style="margin: 0;">
+<body class="elfinder">
 <!-- Element where elFinder will be created (REQUIRED) -->
 <div id="elfinder"></div>