+++ src/resources/views-elfinder/tinymce.php
@@ -62,6 +62,7 @@
         $().ready(function() {
             var elf = $('#elfinder').elfinder({
                 // set your elFinder options here
+                resizable: false,
                 <?php if ($locale) {
     ?>
                     lang: '<?= $locale ?>', // locale
@@ -76,10 +77,16 @@
                 }
             }).elfinder('instance');
         });
+        $(window).resize(function(){
+            var h = ($(window).height());
+            if($('#elfinder').height() != h){
+                $('#elfinder').height(h).resize();
+            }
+        });
     </script>
 
 </head>
-<body>
+<body class="elfinder">
 
 <!-- Element where elFinder will be created (REQUIRED) -->
 <div id="elfinder"></div>