+++ src/resources/views-elfinder/tinymce4.php
@@ -42,6 +42,7 @@
         $().ready(function() {
             var elf = $('#elfinder').elfinder({
                 // set your elFinder options here
+                resizable: false,
                 <?php if ($locale) {
     ?>
                     lang: '<?= $locale ?>', // locale
@@ -56,9 +57,15 @@
                 }
             }).elfinder('instance');
         });
+        $(window).resize(function(){
+            var h = ($(window).height());
+            if($('#elfinder').height() != h){
+                $('#elfinder').height(h).resize();
+            }
+        });
     </script>
 </head>
-<body>
+<body class="elfinder">
     <!-- Element where elFinder will be created (REQUIRED) -->
     <div id="elfinder"></div>
 </body>